....I..... [ZBX-15211] fixed tests compilation, 'cmocka-tests' and 'gen-test-coverage-report' for jenkins

+++ bootstrap.sh
@@ -14,6 +14,8 @@ if [ $# -eq 1 ]; then
 	[ $1 = "tests" ] || usage
 
 	cp -f tests/conf_tests.m4 m4/ || exit $?
+else
+	[ -f m4/conf_tests.m4 ] && rm -f m4/conf_tests.m4
 fi
 
 aclocal -I m4
@@ -22,6 +24,6 @@ autoheader
 automake -a
 automake
 
-if [ -f m4/conf_tests.m4 ]; then
+if [ "x$1" != "xtests" ] && [ -f m4/conf_tests.m4 ]; then
 	rm -f m4/conf_tests.m4
 fi
